Create new postCSS recipe
Hello! I've written up a postCSS recipe, but I've noticed due the project's flow of using useref to handle the minification of the css and my placement on postCSS in this spot, postCSS only runs when useref runs, so this could potentially break the functionality of your css depending on plugin (like with precss).  Is there a recommendation anyone can suggest to rewrite this recipe in a more appropriate and usable manner?  Thanks